requirements.txt
transformers
datasets
sentence-transformers
torch
accelerate
scipy
scikit-learn
nltk
spacy
tqdm
langchain
faiss-cpu
openai
llama-index
unstructured
beautifulsoup4
html2text
regex
python-dateutil
pandas
flask
uvicorn
fastapi
requests
openpyxl
python-docx
python-dotenv
loguru
ipykernel
jupyter
notebook
